fix: remove security check from order init
Some OXID shops have implemented a "one-page-checkout". In this case, the payment ID may only be set when or after the init method of the order controller has been executed.
The Amazon session may also only be created after the init method.
As a security check with regard to AmazonPay, it is sufficient to prevent the order from being completed if there is no valid response from Amazon shortly before the order is completed (-> execute method)