fix: do not set country directly from Amazon response
With Amazon Pay in "PayOnly" mode, Amazon does not send a delivery address.
And even if Amazon sends a delivery address, we still do not need the country ID at this point.
In the method OrderController::setAmazonPayAsPaymentMethod(), the User::getActiveCountry() is called which returns the (delivery) address currently set for the user.