fix: refactore OrderController::execute()
Due to commit
[180d059] "feat: secure order execute from incorrect use"
the parent method (execute) was no longer called and an order with Amazon Pay was never saved.
A redirect to the "amazonPayRedirectUrl" took place, which also resulted in a redirect to the "thankyou-controller".
But without calling parent::execute, there is also no call to Order::finalizeOrder(). In this case we will never save order to OXID database